Move DefaultParser/0 to the main package
